+++ modules/apps/foundation/contacts/contacts-uad/src/main/java/com/liferay/contacts/uad/display/EntryUADDisplay.java
@@ -14,58 +14,16 @@
 
 package com.liferay.contacts.uad.display;
 
-import com.liferay.contacts.model.Entry;
 import com.liferay.contacts.uad.constants.ContactsUADConstants;
-
-import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
-import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
 
 import com.liferay.user.associated.data.display.UADDisplay;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-import java.util.Locale;
-import java.util.Map;
 
 /**
  * @author Brian Wing Shun Chan
- * @generated
  */
 @Component(immediate = true, property =  {
 	"model.class.name=" + ContactsUADConstants.CLASS_NAME_ENTRY}, service = UADDisplay.class)
-public class EntryUADDisplay implements UADDisplay<Entry> {
-	public String getApplicationName() {
-		return ContactsUADConstants.APPLICATION_NAME;
-	}
-
-	public String[] getDisplayFieldNames() {
-		return _entryUADDisplayHelper.getDisplayFieldNames();
-	}
-
-	@Override
-	public String getEditURL(Entry entry,
-		LiferayPortletRequest liferayPortletRequest,
-		LiferayPortletResponse liferayPortletResponse)
-		throws Exception {
-		return _entryUADDisplayHelper.getEntryEditURL(entry,
-			liferayPortletRequest, liferayPortletResponse);
-	}
-
-	public String getKey() {
-		return ContactsUADConstants.CLASS_NAME_ENTRY;
-	}
-
-	@Override
-	public Map<String, Object> getNonanonymizableFieldValues(Entry entry) {
-		return _entryUADDisplayHelper.getUADEntityNonanonymizableFieldValues(entry);
-	}
-
-	@Override
-	public String getTypeName(Locale locale) {
-		return "Entry";
-	}
-
-	@Reference
-	private EntryUADDisplayHelper _entryUADDisplayHelper;
+public class EntryUADDisplay extends BaseEntryUADDisplay {
 }
